Moving JSON entity out of the way and extending structure to recipe table
